closes #421 and #414. Related to #378
