+++ controllers/controllers.js
@@ -1,11 +1,10 @@
 const Controllers = module.exports = {};
 
-Controllers.User = require('./user-controller.js');
-Controllers.Question = require('./question-controller.js');
-Controllers.Resource = require('./resource-controller.js');
-Controllers.Sole = require('./sole-controller.js');
-Controllers.Helper = require('./helper-controller.js');
-Controllers.Dashboard = require('./dashboard-controller.js');
-Controllers.Stats = require('./stats-controller.js');
-Controllers.Test = require('./testing-controller.js');
-Controllers.Admin = require('./admin-controller.js');
+Controllers.User      = require('./user-controller.js');
+Controllers.Question  = require('./question-controller.js');
+Controllers.Resource  = require('./resource-controller.js');
+Controllers.Sole      = require('./sole-controller.js');
+Controllers.Ring      = require('./ring-controller.js');
+Controllers.Stats     = require('./stats-controller.js');
+Controllers.Test      = require('./testing-controller.js');
+Controllers.Admin     = require('./admin-controller.js');
